Update index.md
I changed the URL link to point to Jargon-Busting as the first episode in the "Introduction to Working with Data" Core Lesson listed in the table for the Standard Workshop. I'm not sure if this is correct, but the prior link was linking from this page (lc-overview) to itself (lc-overview).